Change User to Person
Fixed getFirstLast test and method
