+++ app/models/Person.java
@@ -5,12 +5,12 @@ import io.ebean.Finder;
 import javax.persistence.Entity;
 
 @Entity
-public class User extends BaseModel{
+public class Person extends BaseModel{
     public String firstName;
     public String lastName;
     public String email;
 
-    public User (String firstName, String lastName, String email) {
+    public Person(String firstName, String lastName, String email) {
         this.firstName = firstName;
         this.lastName = lastName;
         this.email = email;
@@ -19,12 +19,12 @@ public class User extends BaseModel{
         return this.lastName + ", " + this.firstName;
     }
     public String getFirstLast(){
-        return this.firstName + ", " + this.lastName;
+        return this.firstName + " " + this.lastName;
     }
 
     public String getEmail(){
         return this.email;
     }
 
-    public static Finder<String, User> find = new Finder<>(User.class);
+    public static Finder<String, Person> find = new Finder<>(Person.class);
 }
